backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
pydantic
scikit-learn
